Rename WooPosProductTestUtils
